+++ src/lib/model/arc/arc-type.enum.ts
@@ -1,7 +1,14 @@
 export enum ArcType {
+    REGULAR_PT = 'regular_pt',
+    REGULAR_TP = 'regular_tp',
+    READ = 'read',
+    RESET = 'reset',
+    INHIBITOR = 'inhibitor',
+}
+
+export enum XmlArcType {
     REGULAR = 'regular',
     READ = 'read',
     RESET = 'reset',
     INHIBITOR = 'inhibitor',
-    VARIABLE = 'variable'
 }