+++ src/lib/model/petrinet/place.ts
@@ -1,28 +1,14 @@
 import {I18nString} from '../i18n/i18n-string';
+import {NodeElement} from './node-element';
 
-export class Place {
-    private _id: string;
+export class Place extends NodeElement {
     private _static: boolean;
-    private _x: number;
-    private _y: number;
-    private _label: I18nString;
     private _marking: number;
 
     constructor(x: number, y: number, isStatic: boolean, id: string) {
-        this._id = id;
+        super(id, x, y, new I18nString(''));
         this._static = isStatic;
-        this._x = x;
-        this._y = y;
-        this._label = new I18nString('');
         this._marking = 0;
-    }
-
-    get id(): string {
-        return this._id;
-    }
-
-    set id(value: string) {
-        this._id = value;
     }
 
     get static(): boolean {
@@ -31,30 +17,6 @@ export class Place {
 
     set static(value: boolean) {
         this._static = value;
-    }
-
-    get x(): number {
-        return this._x;
-    }
-
-    set x(value: number) {
-        this._x = value;
-    }
-
-    get y(): number {
-        return this._y;
-    }
-
-    set y(value: number) {
-        this._y = value;
-    }
-
-    get label(): I18nString {
-        return this._label;
-    }
-
-    set label(value: I18nString) {
-        this._label = value;
     }
 
     get marking(): number {
@@ -66,8 +28,8 @@ export class Place {
     }
 
     public clone(): Place {
-        const cloned = new Place(this._x, this._y, this._static, this._id);
-        cloned._label = this._label?.clone();
+        const cloned = new Place(this.x, this.y, this.static, this.id);
+        cloned.label = this.label?.clone();
         cloned._marking = this._marking;
         return cloned;
     }