+++ src/lib/model/transition/transition-event.ts
@@ -26,4 +26,8 @@ export class TransitionEvent extends Event<TransitionEventType> {
         this.postActions.forEach(item => cloned.postActions.push(item.clone()));
         return cloned;
     }
+
+    public isEmpty(): boolean {
+        return super.isEmpty() && this._title.isEmpty();
+    }
 }