+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [2.0.0](https://github.com/netgrif/petriflow.js/releases/tag/v2.0.0) (2023-11-05)
+
+### Added
+- [PF-34] Simulation
+
+### Changed
+- [PF-52] Do not export empty events
+- [PF-54] DataRef order by position
+
+### Fixed
+- [PF-55] Missing data ref attribute - immediate
+
 ## [1.3.5](https://github.com/netgrif/petriflow.js/releases/tag/v1.3.5) (2023-04-13)
 
 ### Fixed